fts: extract reader/sink.rs (TopKEntry, AndSink family, and_heap_push)
Phase 1c (3/3). Move top-k collection — the reversed-order TopKEntry
heap element, drain_top_k_desc, the AndSink trait and its Score/
MustShould/Collect/Count sinks, and the shared and_heap_push — into
reader/sink.rs. Fields/methods pub(super) so core's flat-merge
constructs and drives them. Pure move; FTS + reader suites green.
